unisigned integer for time to supposedly fix the interrupt rotate error, as well as adding a "have to rotate this much" not sure if it actually worked ngl ill test it more. also i would like to test different zigzag numbers. we started at 250 MS, but it didn't really see the back well. 450 was too much (when we're on the edge it kinda just keeps rotating into the edge). we should be able to find something

+++ 500g/Pharmaci/src/Timer.cpp
@@ -1,17 +1,10 @@
 #include "Timer.h"
+#include <Arduino.h>  // for uint32_t on some cores (optional)
 
 bool Timer::getReady() {
-    return (currentTime - previousTime) >= timeInterval;
+    return (uint32_t)(currentTime - previousTime) >= timeInterval;  // rollover-safe
 }
 
-void Timer::setTimeInterval(int time) {
-    timeInterval = time;  
-}
-
-void Timer::setCurrentTime(int time) {
-    currentTime = time;  
-}
-
-void Timer::setPreviousTime(int time) {
-    previousTime = time;
-}
+void Timer::setTimeInterval(uint32_t time) { timeInterval = time; }
+void Timer::setCurrentTime(uint32_t time)   { currentTime = time; }
+void Timer::setPreviousTime(uint32_t time)  { previousTime = time; }
